requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-keycloak>=4.1.0
pydantic>=2.8.0
